refactor: rename active auth module to auth
