+++ frontend/src/features/auth/api/auth.constants.ts
@@ -1,40 +1,23 @@
-export const AUTH_TABS = {
-  SIGNIN: "signin",
-  SIGNUP: "signup",
+export const AUTH_SERVICE_ENDPOINTS = {
+  LOGIN: "/api/auth/login",
+  ME: "/api/auth/me",
+  REFRESH: "/api/auth/refresh",
+  LOGOUT: "/api/auth/logout",
 } as const;
-
-export type AuthTab = (typeof AUTH_TABS)[keyof typeof AUTH_TABS];
 
 export const AUTH_QUERY_PARAMS = {
-  TAB: "tab",
   REDIRECT: "redirect",
-} as const;
-
-export const AUTH_FIELDS = {
-  EMAIL: "email",
-  PASSWORD: "password",
-  FULL_NAME: "fullName",
-  ACCEPT_TERMS: "acceptTerms",
-} as const;
-
-export const AUTH_SERVICE_ENDPOINTS = {
-  ME: "/api/legacy-auth/me",
-  SIGNIN: "/api/legacy-auth/signin",
-  SIGNUP: "/api/legacy-auth/signup",
-  REFRESH: "/api/legacy-auth/refresh",
-  LOGOUT: "/api/legacy-auth/logout",
 } as const;
 
 export const AUTH_ROUTES = {
   ROOT: "/auth",
-  LEGACY: "/auth/legacy",
+  SUCCESS: "/auth/success",
+  FAILURE: "/auth/failure",
   HOME_FALLBACK: "/",
 } as const;
 
 export const HTTP_STATUS = {
   UNAUTHORIZED: 401,
-  BAD_REQUEST: 400,
+  FORBIDDEN: 403,
   NO_CONTENT: 204,
 } as const;
-
-export const ACCESS_TOKEN_EXPIRY_TIME = 1000 * 60 * 30;
