+++ frontend/src/features/auth/index.ts
@@ -1,9 +1,21 @@
 export { AuthProvider, useAuth } from "./model/auth.context";
+export { authKeys } from "./api/auth.query-keys";
+export { meQueryOptions } from "./api/auth.queries";
 export { default as RequireAuth } from "./routing/RequireAuth";
 export { default as RequireGuest } from "./routing/RequireGuest";
-export { default as LegacyAuthPage } from "./ui/LegacyAuthPage";
-export { AUTH_TABS, AUTH_ROUTES } from "./api/auth.constants";
 export {
   buildAuthRoute,
+  consumePostAuthRedirect,
+  getRedirectFromSearchParams,
   getSafeRedirectTarget,
+  persistPostAuthRedirect,
 } from "./routing/auth-redirect";
+export { default as AuthPage } from "./ui/AuthPage";
+export { default as AuthFailurePage } from "./ui/AuthFailurePage";
+export { default as AuthSuccessPage } from "./ui/AuthSuccessPage";
+export { default as AuthLoadingScreen } from "./ui/components/AuthLoadingScreen";
+export {
+  AUTH_QUERY_PARAMS,
+  AUTH_ROUTES,
+  AUTH_SERVICE_ENDPOINTS,
+} from "./api/auth.constants";
