+++ frontend/src/features/auth/routing/RequireAuth.tsx
@@ -1,22 +1,22 @@
 import { Navigate, Outlet, useLocation } from "react-router";
-import { AUTH_TABS } from "../api/auth.constants";
 import { useAuth } from "../model/auth.context";
-import { buildAuthRoute } from "./auth-redirect";
 import AuthLoadingScreen from "../ui/components/AuthLoadingScreen";
+import { buildAuthRoute } from "./auth-redirect";
 
 const RequireAuth = () => {
-  const { status } = useAuth();
+  const { isLoading, status } = useAuth();
   const location = useLocation();
 
-  if (status === "checking") {
+  if (isLoading) {
     return <AuthLoadingScreen />;
   }
 
   if (status === "unauthenticated") {
     const currentPath = `${location.pathname}${location.search}${location.hash}`;
+
     return (
       <Navigate
-        to={buildAuthRoute({ tab: AUTH_TABS.SIGNIN, redirectTo: currentPath })}
+        to={buildAuthRoute({ redirectTo: currentPath })}
         replace
       />
     );
