+++ frontend/src/features/auth/routing/auth-redirect.ts
@@ -1,9 +1,9 @@
 import {
   AUTH_QUERY_PARAMS,
   AUTH_ROUTES,
-  AUTH_TABS,
-  type AuthTab,
 } from "../api/auth.constants";
+
+const POST_AUTH_REDIRECT_STORAGE_KEY = "auth.redirect";
 
 const isSafeInternalRedirect = (
   redirectTarget: string | null | undefined
@@ -21,38 +21,68 @@ export const getSafeRedirectTarget = (
   if (!isSafeInternalRedirect(redirectTarget)) {
     return null;
   }
+
   return redirectTarget;
 };
 
 export const buildAuthRoute = ({
-  tab,
   redirectTo,
 }: {
-  tab: AuthTab;
   redirectTo?: string | null;
 }) => {
-  const params = new URLSearchParams({
-    [AUTH_QUERY_PARAMS.TAB]: tab,
-  });
-
+  const params = new URLSearchParams();
   const safeRedirectTarget = getSafeRedirectTarget(redirectTo);
+
   if (safeRedirectTarget) {
     params.set(AUTH_QUERY_PARAMS.REDIRECT, safeRedirectTarget);
   }
 
-  return `${AUTH_ROUTES.ROOT}?${params.toString()}`;
-};
+  const queryString = params.toString();
 
-export const getAuthTabFromSearchParams = (search: string): AuthTab => {
-  const params = new URLSearchParams(search);
-  const tab = params.get(AUTH_QUERY_PARAMS.TAB);
-  if (tab === AUTH_TABS.SIGNUP) {
-    return AUTH_TABS.SIGNUP;
+  if (!queryString) {
+    return AUTH_ROUTES.ROOT;
   }
-  return AUTH_TABS.SIGNIN;
+
+  return `${AUTH_ROUTES.ROOT}?${queryString}`;
 };
 
 export const getRedirectFromSearchParams = (search: string) => {
   const params = new URLSearchParams(search);
   return getSafeRedirectTarget(params.get(AUTH_QUERY_PARAMS.REDIRECT));
+};
+
+const canUseSessionStorage = () => {
+  return typeof window !== "undefined";
+};
+
+export const persistPostAuthRedirect = (
+  redirectTarget: string | null | undefined
+) => {
+  const safeRedirectTarget = getSafeRedirectTarget(redirectTarget);
+
+  if (!canUseSessionStorage()) {
+    return;
+  }
+
+  if (!safeRedirectTarget) {
+    window.sessionStorage.removeItem(POST_AUTH_REDIRECT_STORAGE_KEY);
+    return;
+  }
+
+  window.sessionStorage.setItem(
+    POST_AUTH_REDIRECT_STORAGE_KEY,
+    safeRedirectTarget
+  );
+};
+
+export const consumePostAuthRedirect = () => {
+  if (!canUseSessionStorage()) {
+    return null;
+  }
+
+  const redirectTarget = window.sessionStorage.getItem(
+    POST_AUTH_REDIRECT_STORAGE_KEY
+  );
+  window.sessionStorage.removeItem(POST_AUTH_REDIRECT_STORAGE_KEY);
+  return getSafeRedirectTarget(redirectTarget);
 };
