+++ frontend/src/features/auth/ui/AuthFailurePage.tsx
@@ -1,5 +1,5 @@
 import { Link } from "react-router";
-import { AUTH_V2_ROUTES } from "../api/auth.constants";
+import { AUTH_ROUTES } from "../api/auth.constants";
 import AuthOutcomePage from "./AuthOutcomePage";
 
 const AuthFailurePage = () => {
@@ -10,7 +10,7 @@ const AuthFailurePage = () => {
       footer={
         <div className="flex items-center gap-3">
           <Link
-            to={AUTH_V2_ROUTES.ROOT}
+            to={AUTH_ROUTES.ROOT}
             className="inline-flex items-center justify-center border border-zinc-300 bg-white px-3 py-2 text-sm text-black transition-colors hover:bg-zinc-50"
           >
             Back to sign in
