+++ frontend/src/features/auth_legacy/api/auth.query-keys.ts
@@ -1,0 +1,4 @@
+export const authKeys = {
+  all: ["auth"] as const,
+  me: () => [...authKeys.all, "me"] as const,
+};
