+++ frontend/src/features/auth_legacy/index.ts
@@ -1,0 +1,9 @@
+export { AuthProvider, useAuth } from "./model/auth.context";
+export { default as RequireAuth } from "./routing/RequireAuth";
+export { default as RequireGuest } from "./routing/RequireGuest";
+export { default as LegacyAuthPage } from "./ui/LegacyAuthPage";
+export { AUTH_TABS, AUTH_ROUTES } from "./api/auth.constants";
+export {
+  buildAuthRoute,
+  getSafeRedirectTarget,
+} from "./routing/auth-redirect";
