+++ frontend/src/features/auth_legacy/routing/RequireGuest.tsx
@@ -1,23 +1,28 @@
 import { Navigate, Outlet, useLocation } from "react-router";
 import { AUTH_ROUTES } from "../api/auth.constants";
 import { useAuth } from "../model/auth.context";
-import AuthLoadingScreen from "../ui/components/AuthLoadingScreen";
 import { getRedirectFromSearchParams } from "./auth-redirect";
+import AuthLoadingScreen from "../ui/components/AuthLoadingScreen";
 
 const RequireGuest = () => {
-  const { isLoading, status } = useAuth();
+  const { status } = useAuth();
   const location = useLocation();
 
-  if (isLoading) {
+  if (status === "checking") {
     return <AuthLoadingScreen />;
   }
 
   if (status === "authenticated") {
-    const redirectTarget =
-      getRedirectFromSearchParams(location.search) ??
-      AUTH_ROUTES.HOME_FALLBACK;
+    const redirectTarget = getRedirectFromSearchParams(
+      location.search
+    );
 
-    return <Navigate to={redirectTarget} replace />;
+    return (
+      <Navigate
+        to={redirectTarget ?? AUTH_ROUTES.HOME_FALLBACK}
+        replace
+      />
+    );
   }
 
   return <Outlet />;
