+++ frontend/src/features/auth_legacy/ui/LegacyAuthPage.tsx
