+++ frontend/src/features/auth_legacy/ui/components/PasswordInput.tsx
