+++ frontend/src/features/auth_legacy/ui/hooks/usePasswordValidation.ts
