+++ frontend/src/features/auth_legacy/ui/utils/validation.ts
