--- frontend/src/features/auth_v2/api/auth.query-keys.ts
@@ -1,4 +1,0 @@
-export const authV2Keys = {
-  all: ["auth_v2"] as const,
-  me: () => [...authV2Keys.all, "me"] as const,
-};
