--- frontend/src/features/auth_v2/index.ts
@@ -1,21 +1,0 @@
-export { AuthProvider, useAuth } from "./model/auth.context";
-export { authV2Keys } from "./api/auth.query-keys";
-export { meQueryOptions } from "./api/auth.queries";
-export { default as RequireAuth } from "./routing/RequireAuth";
-export { default as RequireGuest } from "./routing/RequireGuest";
-export {
-  buildAuthRoute,
-  consumePostAuthRedirect,
-  getRedirectFromSearchParams,
-  getSafeRedirectTarget,
-  persistPostAuthRedirect,
-} from "./routing/auth-redirect";
-export { default as AuthPage } from "./ui/AuthPage";
-export { default as AuthFailurePage } from "./ui/AuthFailurePage";
-export { default as AuthSuccessPage } from "./ui/AuthSuccessPage";
-export { default as AuthLoadingScreen } from "./ui/components/AuthLoadingScreen";
-export {
-  AUTH_V2_QUERY_PARAMS,
-  AUTH_V2_ROUTES,
-  AUTH_V2_SERVICE_ENDPOINTS,
-} from "./api/auth.constants";
