--- frontend/src/pages/Login.tsx
@@ -1,118 +1,0 @@
-import { useState, type FormEvent } from "react";
-import { Button } from "@/components/ui/button";
-import {
-  Card,
-  CardContent,
-  CardDescription,
-  CardFooter,
-  CardHeader,
-  CardTitle,
-} from "@/components/ui/card";
-import { AUTH_SERVICE_ENDPOINTS } from "@/features/auth/api/auth.constants";
-import { Input } from "@/components/ui/input";
-import { httpClient } from "@/lib/httpClient";
-import { useNavigate } from "react-router";
-
-interface LoginCredentials {
-  email: string;
-  password: string;
-}
-
-interface LoginResponse {
-  token?: string;
-  message?: string;
-}
-
-export default function Login() {
-  const [email, setEmail] = useState<string>("");
-  const [password, setPassword] = useState<string>("");
-  const [isLoading, setIsLoading] = useState<boolean>(false);
-  const [error, setError] = useState<string>("");
-  const navigate = useNavigate();
-
-  const handleSubmit = async (event: FormEvent<HTMLFormElement>) => {
-    event.preventDefault();
-    setError("");
-    setIsLoading(true);
-
-    try {
-      const credentials: LoginCredentials = { email, password };
-      const response = await httpClient.post<LoginResponse>(
-        AUTH_SERVICE_ENDPOINTS.SIGNIN,
-        credentials
-      );
-
-      // Handle successful login
-      console.log("Login successful:", response);
-      navigate("/");
-      // Navigate to dashboard or handle token storage here
-    } catch (err) {
-      const errorMessage =
-        err instanceof Error
-          ? err.message
-          : "Login failed. Please try again.";
-      setError(errorMessage);
-    } finally {
-      setIsLoading(false);
-    }
-  };
-
-  return (
-    <div className="flex flex-col items-center justify-center h-screen">
-      <Card className="w-full max-w-sm">
-        <CardHeader>
-          <CardTitle>Login to your account</CardTitle>
-          <CardDescription>
-            Enter your credentials to access your account
-          </CardDescription>
-        </CardHeader>
-        <CardContent>
-          <form onSubmit={handleSubmit}>
-            <div className="flex flex-col gap-4">
-              <div className="grid gap-2">
-                <label htmlFor="email" className="text-sm font-medium">
-                  Email
-                </label>
-                <Input
-                  id="email"
-                  type="email"
-                  placeholder="m@example.com"
-                  value={email}
-                  onChange={(e) => setEmail(e.target.value)}
-                  required
-                  disabled={isLoading}
-                />
-              </div>
-              <div className="grid gap-2">
-                <label
-                  htmlFor="password"
-                  className="text-sm font-medium"
-                >
-                  Password
-                </label>
-                <Input
-                  id="password"
-                  type="password"
-                  value={password}
-                  onChange={(e) => setPassword(e.target.value)}
-                  required
-                  disabled={isLoading}
-                />
-              </div>
-              {error && <p className="text-sm text-red-600">{error}</p>}
-            </div>
-            <CardFooter className="px-0 pt-6">
-              <Button
-                type="submit"
-                className="w-full"
-                disabled={isLoading}
-              >
-                {isLoading ? "Logging in..." : "Login"}
-              </Button>
-            </CardFooter>
-          </form>
-        </CardContent>
-      </Card>
-    </div>
-  );
-}
